students/pyproject.toml
[tool.poetry]
name = "students"
version = "0.1.0"
description = ""
authors = ["Yartsev Vladimir <v.rainovsky@yandex.ru>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
strawberry-graphql = {extras = ["fastapi"], version = "^0.134.3"}
uvicorn = "^0.18.3"
sqlalchemy = {extras = ["aiosqlite"], version = "^1.4.41"}
taskipy = "^1.10.3"
loguru = "^0.6.0"


[tool.taskipy.tasks]
server = "uvicorn server:app --host 0.0.0.0 --port 11801 --reload"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
